Added margin to sprite classes
Added margin properties to the CSS classes 'sprite64', 'sprite16', and 'icon'. These margin adjustments are intended to enhance the spacing and alignment of the corresponding elements within the dashboard layout.

Details:
- The 'sprite64' class now includes a top margin of 10px to introduce vertical spacing. This change will be beneficial for the eventual addition of a "Party" section to the bot.
- Both the 'sprite16' and 'icon' classes have been given a left margin of 5px to create horizontal spacing.

The margin additions have been done for an improved visual presentation and a better positioning of the sprites and icons in the overall design. Your feedback on these changes is welcome.